+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./bimark";
 export * from "./parser";
 export * from "./bidoc";
 export * from "./model";
+export * from "./biml";